fix/ moved button style and test routes to seperate files, removed commented parts of the code
