+++ lib/features/home/views/home_view.dart
@@ -1,52 +1,7 @@
 import "package:flutter/material.dart";
 import "../../../common/widgets/section_header.dart";
-// import "package:flutter_riverpod/flutter_riverpod.dart";
-// import "../../../app/app.dart";
-// import "../../route_map/repository/route_map_repository.dart";
-
-// 1. Model trasy
-class RouteItem {
-  final String id;
-  final String name;
-  final String imageUrl;
-  final String details;
-
-  RouteItem({required this.id, required this.name, required this.imageUrl, required this.details});
-}
-
-// 2. Dane testowe
-final List<RouteItem> testRoutes = [
-  RouteItem(
-    id: "1",
-    name: "Spacer po Starym Mieście",
-    imageUrl: "https://picsum.photos/seed/oldtown/300/200",
-    details: "pierwsza trasa",
-  ),
-  RouteItem(
-    id: "2",
-    name: "Zielona Dolina",
-    imageUrl: "https://picsum.photos/seed/valley/300/200",
-    details: "druga trasa",
-  ),
-  RouteItem(
-    id: "3",
-    name: "Wzgórza Panorama",
-    imageUrl: "https://picsum.photos/seed/hills/300/200",
-    details: "trzecia trasa",
-  ),
-  RouteItem(
-    id: "4",
-    name: "Nadbrzeżny Szlak",
-    imageUrl: "https://picsum.photos/seed/coast/300/200",
-    details: "czwarta trasa",
-  ),
-  RouteItem(
-    id: "5",
-    name: "Leśna Ścieżka",
-    imageUrl: "https://picsum.photos/seed/forest/300/200",
-    details: "piąta trasa",
-  ),
-];
+import "./home_view_styles.dart";
+import "./test_routes_list.dart";
 
 class MyHomeView extends StatefulWidget {
   const MyHomeView({super.key, required this.title});
@@ -94,12 +49,7 @@ class _MyHomeViewState extends State<MyHomeView> {
                   return SizedBox(
                     width: 150,
                     child: ElevatedButton(
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.white,
-                        elevation: 2,
-                        padding: EdgeInsets.zero,
-                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-                      ),
+                      style: sharedCardButtonStyle(minimumSize: 120, zeroPadding: true),
                       onPressed: () {
                         // Navigator.push lub showDialog z route.id
                       },
@@ -108,7 +58,13 @@ class _MyHomeViewState extends State<MyHomeView> {
                         child: Stack(
                           fit: StackFit.expand,
                           children: [
-                            Image.network(route.imageUrl, fit: BoxFit.cover),
+                            Image.network(
+                              route.imageUrl,
+                              fit: BoxFit.cover,
+                              errorBuilder: (context, error, stackTrace) {
+                                return const Center(child: Icon(Icons.broken_image, size: 48, color: Colors.grey));
+                              },
+                            ),
                             // półprzezroczysty overlay i nazwa
                             Container(
                               alignment: Alignment.bottomCenter,
@@ -147,12 +103,7 @@ class _MyHomeViewState extends State<MyHomeView> {
 
             // Przycisk rozpocznij swój spacer – biały card z zielonym accentem
             ElevatedButton(
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.white,
-                elevation: 2,
-                minimumSize: const Size(double.infinity, 100),
-                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-              ),
+              style: sharedCardButtonStyle(),
               onPressed: () {},
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -166,12 +117,15 @@ class _MyHomeViewState extends State<MyHomeView> {
                       child: Icon(Icons.explore, color: Colors.white, size: 72),
                     ),
                   ),
-                  const SizedBox(
-                    width: 100,
-                    child: Text(
-                      "Rozpocznij swój spacer",
-                      style: TextStyle(fontSize: 16, color: Colors.black),
-                      textAlign: TextAlign.end,
+
+                  Flexible(
+                    child: ConstrainedBox(
+                      constraints: const BoxConstraints(maxWidth: 120),
+                      child: const Text(
+                        "Rozpocznij swój spacer",
+                        style: TextStyle(fontSize: 16, color: Colors.black),
+                        textAlign: TextAlign.end,
+                      ),
                     ),
                   ),
                 ],
@@ -186,12 +140,7 @@ class _MyHomeViewState extends State<MyHomeView> {
                 children: [
                   Expanded(
                     child: ElevatedButton(
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.white,
-                        elevation: 2,
-                        minimumSize: const Size(double.infinity, 80),
-                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-                      ),
+                      style: sharedCardButtonStyle(minimumSize: 80),
                       onPressed: () {},
                       child: const Column(
                         // tu Column, bo ikona nad tekstem
@@ -213,12 +162,7 @@ class _MyHomeViewState extends State<MyHomeView> {
                   const SizedBox(width: 32),
                   Expanded(
                     child: ElevatedButton(
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.white,
-                        elevation: 2,
-                        minimumSize: const Size(double.infinity, 80),
-                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-                      ),
+                      style: sharedCardButtonStyle(minimumSize: 80),
                       onPressed: () {},
                       child: const Column(
                         mainAxisSize: MainAxisSize.min,
@@ -245,29 +189,3 @@ class _MyHomeViewState extends State<MyHomeView> {
     );
   }
 }
-
-// class TestProviderWidget extends ConsumerWidget {
-//   const TestProviderWidget({super.key});
-
-//   @override
-//   Widget build(BuildContext context, WidgetRef ref) {
-//     final routes = ref.watch(fetchAllRoutesProvider);
-//     final route_2 = ref.watch(fetchRouteWithIdProvider(2));
-
-//     return Column(
-//       mainAxisAlignment: MainAxisAlignment.center,
-//       children: <Widget>[
-//         routes.when(
-//           data: (routes) => Text("Routes:\n$routes"),
-//           loading: CircularProgressIndicator.new,
-//           error: (e, _) => Text("Error: $e"),
-//         ),
-//         route_2.when(
-//           data: (route) => Text("\nRoute 2:\n$route"),
-//           loading: CircularProgressIndicator.new,
-//           error: (e, _) => Text("Error: $e"),
-//         ),
-//       ],
-//     );
-//   }
-// }
